Fits a local regression model with dispersion as a function of total gene counts. Predicted values are the expected dispersion for genes with similar expression levels The initial dispersion estimates are shrunk towards the common dispersion

Fits a local regression model with dispersion as a function of total gene counts. Predicted values are the expected dispersion for genes with similar expression levels The initial dispersion estimates are shrunk towards the common dispersion

correct_theta_bygroup(
  estimates,
  delta_set = 50,
  N_set = 30,
  thetaFilter = NULL
)

Arguments

estimates

Output of estim_bbparams_bygroup

delta_set

Delta parameter

N_set

N parameter

thetaFilter

Minimum dispersion value, genes with dispersion below the set threshold are excluded from the shrinking procedure.

Examples

correct_theta()
#> Error in eval(assertion, env): argument "estimates" is missing, with no default