Estimate beta-binomial distribution parameters per gene. The users can remove low abundant or low expressed genes through quality control parameters min_counts and min_cells.

Estimate beta-binomial distribution parameters per gene. The users can remove low abundant or low expressed genes through quality control parameters min_counts and min_cells.

Usage

estim_bbparams(
  a1_counts,
  tot_counts,
  min_counts = 0,
  min_cells = 5,
  cores = NULL
)

Arguments

a1_counts

Integer matrix (genes x cells): allele-1 counts.

tot_counts

Integer matrix (genes x cells): total counts (same dimenstions and rownames as a1_counts).

min_counts

Integer >= 0. Minimum reads per cell to include (default 0). Cells with a number of mapped reads less than min_counts are excluded from the estimation

min_cells

Integer >= 1. Minimum number of cells per gene to fit (default 5). Genes with a number of cells less than min_cells are excluded from the estimation.

cores

number of processing cores (default 1).

Examples

estim_bbparams(a1_counts,
              tot_counts,
              min_counts = 0,
              min_cells = 5,
              cores = NULL)
#> Error: object 'a1_counts' not found