posterior counts simulated under the expected dispersion

posterior counts simulated under the expected dispersion

Usage

make_plotdf_simul(
  a1_counts,
  tot_counts,
  gene,
  metadata,
  split.var = "group",
  order.by = "time",
  estimates_group,
  init.seed = 10001101
)

Arguments

a1_counts

Reference allele counts

tot_counts

Total gene counts counts

gene

gene symbol or other id, the same style that is used for the row names in the count matrices

metadata

Metadata object containing cell level information. Metadata rownames must match the column names of the count matrices.

split.var

Name of the variable which will be used to split the cells.

order.by

variable to order allelic ratio by

estimates_group

output of estim_bbparams_bygroup()

init.seed

set seed value

Examples

make_plotdf_simul()
#> Error in make_plotdf_simul(): argument "a1_counts" is missing, with no default